Fix bimpc distance-calc pipeline, wire up ICP result routing, refresh build
Root cause of the bimpc "Buffer overflow" was a float/double mismatch in
BimPcChunk/MeshChunk bounding-box (de)serialization, not a network issue.
processBimPc() now computes real point-to-mesh distances via KD-tree instead
of returning dummy values. Fixed the whole solution's build (Test/* projects
were silently stale) by force-enabling the v140 gtest package under the
available v145 toolset, and hardened ImportPointclouds2 against indexing
past a failed load. Corrected a result_type dispatch bug and implemented the
long-TODO handleIcpResult, while documenting that Master-only synchronous
ICP (not slave-distributed fine alignment) remains the supported path. Made
Slave processing_threads actually configurable end-to-end (it was silently
dropped between the Agent and SlaveApp.exe). Replaced the stale
README/CMakeLists/Makefile (still describing a src/master,src/slave layout
that no longer exists) with accurate current-structure docs. Verified the
whole pipeline against the full 3.4GB samsung.las dataset: 1,294 chunks,
125M points, 100% success. Full findings and verification steps are in
CHANGELOG_2026-07-11.md.

+++ src/NetworkManager.cpp
@@ -474,8 +474,12 @@ namespace DPApp {
             }
 
             /// 3. Bounding box
-            meshChunk.min_x = reader.read<float>(); meshChunk.min_y = reader.read<float>(); meshChunk.min_z = reader.read<float>();
-            meshChunk.max_x = reader.read<float>(); meshChunk.max_y = reader.read<float>(); meshChunk.max_z = reader.read<float>();
+            /// [Fix] MeshChunk::min_x..max_z are double (see include/bim/MeshChunk.h), but this
+            /// used to read them as float, under-consuming 24 bytes and desynchronizing every
+            /// read that follows in the enclosing BimPcChunk (garbage bim_data_size -> "Buffer
+            /// overflow: insufficient data to read." when deserializing distributed BIM_PC2_DIST tasks).
+            meshChunk.min_x = reader.read<double>(); meshChunk.min_y = reader.read<double>(); meshChunk.min_z = reader.read<double>();
+            meshChunk.max_x = reader.read<double>(); meshChunk.max_y = reader.read<double>(); meshChunk.max_z = reader.read<double>();
 
             return meshChunk;
         }
@@ -549,8 +553,12 @@ namespace DPApp {
             }
 
             /// 3. Bounding Box
-            chunk.min_x = reader.read<float>(); chunk.min_y = reader.read<float>(); chunk.min_z = reader.read<float>();
-            chunk.max_x = reader.read<float>(); chunk.max_y = reader.read<float>(); chunk.max_z = reader.read<float>();
+            /// [Fix] BimPcChunk::min_x..max_z are double (see include/PointCloudTypes.h), but this
+            /// used to read them as float, under-consuming 24 bytes. Every subsequent read (starting
+            /// with bim_data_size below) was then offset into the wrong bytes, producing a garbage
+            /// size and triggering "Buffer overflow: insufficient data to read." on the Slave side.
+            chunk.min_x = reader.read<double>(); chunk.min_y = reader.read<double>(); chunk.min_z = reader.read<double>();
+            chunk.max_x = reader.read<double>(); chunk.max_y = reader.read<double>(); chunk.max_z = reader.read<double>();
 
             /// 4. MeshChunk (Nested deserialization)
             uint32_t bim_data_size = reader.read<uint32_t>();
